Add Interrupts + Implement HALTing
That includes the halt bug which was a complete horror to implement! Nonetheless I think I've done it right! :smiling_face_with_tear:
Sad thing is that I won't know until testing, but that is a problem for later :weary:
For now tho I think I'll just finish the rest of the instructions in a single commit as they seem extremely trivial, and no more difficult ones left :breath_out:
Also, hi stranger reading this, how are you? :grin: